Generic vector device - rework optimisation slightly
Spotted while working on bug #695647, the xpswrite output from PDF file
CATX4547.pdf.

I added an optimisation to avoid writing a path entirely (avoiding a
beginpath/endpath) when we wouldn't write anything in between the begin
and end.

There was an error in that optimisation, it should only be applied when
prect is assigned to list->single, not generally. That was causing an
image not to be written because (yuck) the code reuses the writing of
the clip in order to write the area of the image.

This shows no differences on the cluster, surprisingly.