xpswrite - 3 clipping fixes
Bug #695647 "Differences in xpswrite"

The first fix is a typo in a previous commit where I accidentally left
a variable in a gs_snprintf call that should not have been present (the
'colour' when writing a clip...).

Secondly, when ending a path which is not a fill, stroke or image fill
we need to explicitly close the path.

Finally; because the xpswrite device doesn't write the clip as the
property of a new canvas, we need to apply it to each path (fill, stroke
or image fill), even if we have previously written it for a different
path. The problem is that the vector device uses the device's stored
clip_path_id, and checks it against the pcpath being used. If they are
the same, it doesn't write the clip. Fix that by explicitly resetting
the clip path ID. If we ever improve this by creating a sub-canvas with
a clip property this can be done away with.