PXLcolor device- don't write CIE images as if they were RGB
Bug #694741 "Differences in output via pxlcolor vs. direct"

The input PostScript file, unusually, contains a number of images which
use a CIEBased colour space (mostly CIEBasedABC). The majority of these
images are skewed or rotated, which means that the PCLcolor device is
unable to emit them as images and so decomposes them into trapezoids
and writes those.

These images render correctly.

Some few images, however, are simple horizontal or vertical oriented
images, and in this case the PXLcolor device decides that they can be
written as images. The code does spot the use of a number of different
colour spaces which it cannot support (eg ICCBased) and decomposes those
to trapezoids, but it did not spot the use of CIEBased colour.

This resulted in the CIE image samples being treated as RGB, which leads
to the colour being incorrect.

This commit just adds the CIEBased range of spaces to the ones that the
device knows it can't handle, and instead turns into trapezoids, which
works.