Bug 706788: Fix interaction of devn code with tags.
This bug has been a voyage of discovery. First off, update the
gdevpsd device with some additional comments to note the
odd and confusing setup of the devn structure that had me
running down several blind alleys.

We send a larger list of colorant names than the standard code
ever uses - just so we can conveniently get to the rest of the
list ourselves in the psd output routines.

The actual bug is due to the devn code not understanding tags.
It assumes that every component between "num_std_colorant_names"
and "max_components" is available for use as a spot color.

The fix implemented here is to extend the devn_params code with
a 'num_reserved_colorants' field, and to not allow the auto
code to expand into this. We set that to 1 in the PSD device.