Colour management - don't apply object tagging parameters to CIE colours
Bug #706789 "Problems with CIEBased colour space"

The problem is that CIEBased colour spaces end up with an ICC profile
which (I think) has an incorrect 'data_cs' value. For example
CIEBasedABC has a data_cs of gsRGB, which it clearly is not.

I don't want to try and change that, though, for two reasons; firstly
the enumerator containing gsRGB doesn't have values for CIEBasedA and
friends. Secondly because I suspect that this assumption may be baked
into the colour management in other places and altering it could lead
to a lengthy bug tail.

Fortunately, the profiles generated for CIEBased colour spaces have an
additional member set 'default_match' and we can check that as well.

In this commit, if the profile has a 'default_match' which lies in the
CIEA->CRD (Color Rendering Dictionary) range then we do not apply the
functionality of the tagged colour management to it. That is now only
applied to colours specified in the Device spaces.