Bug 706793(2): Build and reuse a dictionary of the TTF post table
Previously we relied on lazily evaluating the contents of the TTF post table
which, by its nature, means a fairly inefficient process.

But for very large TTF fonts, that becomes a performance bottleneck, especially
with pdfwrite because it evaluates every glyph in the font.

So create a pdfi dictionary and use that to lookup entries.

This is made more worthwhile because of the prvious commit which allows for the
pdfi objects to be shared between the font instances.