graphics library - initialise transfer maps
Bug #706794 "Global buffer overflow read of size 2 in gx_devn_reduce_colored_halftone (gxdevndi.c:179)"

The test file, when decoded from binary tokens/sequences, performs this:

{8 1 1 matrix 0 string image} settransfer

So that attempts to draw an image inside the transfer function. Because
we are rendering to a 1-bit device we attempt to use the transfer
function we are in the middle of defining to remap the 0 and 1 values
of the image data.

If we are lucky that doesn't cause a problem, but... The transfer map
values are stored as unsigned short (frac), but are only expected to
have positive values (0->32767 = 0->1). Because the map values are not
initialised they may take on illegal values. Later on we apply these,
then try to use the result to look up a table. Because the table is set
up only to handle positive values, we can end up overrunning it.

To avoid this, set the transfer map values explicitly to zero when we
create the map (before we populate it). This makes certain that we can't
use uninitialised data and is at least always valid. This is very
slightly slower, but evaluating a transfer function in PostScript is
already slow enough that we won't worry about it.