Graphics library - fix buffer under-read in threshold arrays
Bug #706795 "Heap buffer overflow read of size 264 in fill_threshold_buffer (gxht_thresh.c:774)"

This one is complicated to explain. When applying a threshold array it
can be the case that the area doesn't precisely butt up to the left
edge of the object we are drawing. So we need to use a 'tile' from the
array and apply just that to the object.

In order to generate that tile we (obviously) have to offset into the
threshold array. The code as it stands just applies the 'left offset'
without considering whether there is a valid memory area to copy from.

If we are trying to use the initial data in the threshold array this
can mean that the starting point (taking the left offset into account)
is before the start of the actual threshold array, leading to us trying
to read from an uninitialised (and potentially non-existent) area of
memory.

In point of fact we won't be using the data we are reading here, it's
only needed so that we generate the tile correctly with an offset into
the main array, but reading prior to the beginning of the buffer is not
a good idea.

This commit simply checks for this condition and, if it should occur,
just moves the pointers along as if the data had been read but without
actually reading it. Since we won't be using this data it doesn't
matter that we are leaving the destination uninitialised in this case.