Improvement to font object reuse/cache
Instead of storing a font object as a non-standard entry in a FontDescriptor
object, implement a font object cache, which matches a object number from a
FontDescriptor with an existing font object.

This is a simple, unsorted array of structs since the number of entries
involved is generally small.

Currently, the cache is created and cleared per page.