pdfwrite - remove restrictions on document info strings
Bug #706781 "German Umlauts in Info Dict not possible when generating PDF/A"

It appears that recent versions of VeraPDF conform to what I believe the
specification for PDF/A says, and permit the Document Information
Dictionary strings and the corresponding XMP data to be *equivalent*,
not *identical*.

In the past we have observed that PDF validators (including earlier
versions of VeraPDF) expected the strings to be exactly the same. So
any strings encoded using PDFDocEncoding and using character codes
outside the ASCII range would fail to validate.

This may still be the case for other validators but I'm only going to
consider VeraPDF because it is a decent validator and reputable.

This commit removes the limit on using UTF-16BE encoded strings, and
PDFDocEncoding strings with character codes outside the ASCII range.
This has meant adding some extra validation to PDFDocEncoding strings
which apply in pdf_xmp_write_translated. We now check all the character
codes in a PDFDocEncoding string and throw an error on the use of
undefined character codes.

Stefan Ziegler <sz5000@gmx.de> provided the original patch, but I chose
to diverge from that significantly.