oss-fuzz 60203/60204/60207: Fix mistakes in glyph_lenths table copying
Two mistakes: I used TrueNumGlyphs when it should have been numGlyphs - the
former is the value read from the font, the latter is the value derived from
the size of the loca table, and loca table is where the glyph_lenths table is
derived from.

Secondly, the glyph_lengths table is actually one entry larger than the loca
table.