Bug 706811: Make Postscript error handlers safer
Call the internally defined method for printing, rather than the language
defined one ("=only" as opposed to "=") since the latter can be redefined.

Make the internal one an operator, so it gets bound into the procedure at
definition.

Thanks for Alan Goodman for the fuzz testing which uncovered this bug.