Graphics library - sanity check dash pattern offset
Bug #706812 "Heap buffer overflow read of size 4 in gx_set_dash (gsline.c:204)"

The problem is that the dash offset is *huge* 6.6x10^30 or so. This is
then stored in a floating point number, and more calculations performed
which, unfortunately result in accuracy compromises.

The end result is that we end up with a large number for the offset
length into the dash, larger than the actual length of the dash pattern
which means we end up indexing off the pattern dash array.

We could, possibly, solve this by using doubles instead of floats but
I think sufficiently large numbers would expose the problem again.
Instead just make sure that, after we've calculate the offset along the
dash length, it does not exceed the length of the pattern, that way we
will not index off the pattern array.

Thanks to Alan Goodman for the fuzz testing which uncovered this bug.