Graphics library - don't use an uninitialised threshold array
Bug #706817 "Segfault on unknown address in fill_threshold_buffer (gxht_thresh.c:788)"

The (decoded) file does something like:

4 null 8 matrix {b0 (0000000000)} image

When the image operator calls the procedure to get data, it executes
'b0' which implicitly calls initgraphics, which resets the halftone.

When we then try to actually draw the image samples, we try to use a
halftone which has not now been initialised, leading to a crash.

Simply check the halftone component threshold to see if it is non-NULL
before trying to use it.

Thanks to Alan Goodman for the fuzz testing which uncovered this bug.