Bug 706819: images: NULL b_tile pointer before returning to interpreter
When rendering a monochrome image to a halftoned device, we cache colour
values in the image_clues. Those are device colours, which can include the
halftone and a pointer to the halftone tile.

The tile (binary.b_tile) is just a pointer to the tile in the tile cache, which
is actually "owned" by the halftone in the graphics state. So if the halftone
in the graphics state changes, the pointer becomes invalid.

Whilst this should not happen, it can if the Postscript image source is a
procedure, and procedure calls setscreen (it shouldn't, but nothing says it
cannot).

So NULL those pointers in the image clues before returning to the interpreter
for more data.

Other than the loop to do the NULLing, this doesn't affect performance, since
it is NULLed for each new sample value anyway.

Thanks to Alan Goodman for the fuzz testing which uncovered this bug.