Graphics library - don't permit CCITTFaxDecode to have /Columns 0
Bug 706816 "The program seems to be stuck in a loop"

The PDF file has a /Filter of CCITTFaxDecode with a Params dictionary
which contains a /Columns value of 0. Clearly this is nonsense.

This commit checks the value of /Columns and, if it is less than or
equal to 0, throws an error.