Fix Nup SEGVs.
When a devn device is called to put_params, we watch for the color
setup changing. If it changes, we reset it and close the device (so
the device closes with the "old" setup), and the put the color setup
back to the "new" state so it can be opened again in future.

We were not testing 'num_planar_planes' as part of this setup,
when we should have been. Extend the support for this.

This was showing up as crashes with Nup due to the Nup device
triggering a flush on the close.