Graphics library - avoid divide by zero with ridiculous resolutions
Bug #706847 "FPE gs_screen_enum_init_memory / zscreen_enum_init"

The specimen file consists solely of:

<< /HWResolution [32 65535] >> setpagedevice

Which, while legal, is plainly ridiculous. The code for selecting a
halftone tile wasn't checking for a potential divide by zero (caused
by the resolution).

So check for that and just continue if we would get a divide by zero.
Either we'll find a better match or, for silly cases like this, end
up using the default. The result will be awful no matter what, in this
case.