Transparency device - check saved function before use.
OSS-fuzz #60427

The file is, of course, corrupted beyond repair. However the PDF14
device obviously shouldn't crash. I don't know how it is getting into
this state, but the most obvious fix is imply to ensure that the
function we're going to call isn't NULL before we try and execute it.

If it is, we just return -1 as an error (obviously taking inspiration
from the PCL interpreter here).