Bug 706828: Fix Line buffer overflow in PWGDecode filter.
The PWGDecode filter saves the number of bytes to copy
or fill in the stream state before validating its value.
Even if the validation ends with an error, it is possible
to invoke another read, which will continue with the
updated state.

Credit (and thanks!) to Zdenek Hutyra (zhutyra@centrum.cz)
for the bug report and the patch to fix it.