PostScript interpreter - revise use of ref_stack_index()
Bug #706837 "Segfault on unknown address in real_param (iutil.c:861)"

The problem here is that ref_stack_index() returns NULL if it cannot
retrieve the requested object from the nominated stack. It turns out
that there are a *lot* of places where we do not test that and simply
assume that we will always get a valid object returned.

Many of those places then dereference the object, on the assumption
that it is valid. An exception is dict_find(), which does check the
supplied dictionary, so those uses are unchanged.

This (unfortunately extensive) commit visits all the places where we
use ref_stack_index() and checks to see if the returned value is
either checked, or passed to a function which checks it. Anywhere which
does neither is revised to return a stackunderflow error, or take
similar preventative action (such as ignoring the object instead of
processing it).

Thanks to Alan Goodman for the fuzz testing which uncovered this bug.