PDF interpreter - don't use PS string as C string
Bug #706839 "Segfault on unknown address in pagelist_parse_to_array (pagelist.c:45)"

PostScript strings aren't C strings, in particular they may not be
NULL-terminated. In this case the string was 0 length so the pointer
was NULL and we tried to use it as a C string, but the principle is the
same; we can't use PostScript strings as C strings.

Now that we are not treating the PageList PostScript string as if it
were a C string, we don't need to NULL terminate the PostScript string,
so get rid of that.

Thanks to Alan Goodman for the fuzz testing which uncovered this bug.