Graphics library - sanity check number of inputs for sampled function
Bug #706856 "Heap buffer overflow write of size 1 in sampled_data_continue (zfsample.c:532) "

The problem exhibits far from this point, but is due to the number of
components in a DeviceN space being too large to execute the sampled
function to sample the space.

Check the number of components early to avoid overrrunning a buffer.

Thanks to Alan Goodman for the fuzz testing which uncovered this bug.