PostScript interpreter - rework lookup procedure data sources
Bug #706845 "Segfault on unknown address in float_params (iutil.c)"

The basic problem here is very similar to a number of previous reports;
the procedure used to build the indexed colour lookup does a 'grestore'
as the first action. Since there was no previous gsave this restores
back one level of graphics state from the current one.

Obviously this is silly; but it appears to be legal and we should be
able to cater for it. NB similar problems arose with transfer functions
previously which made used marking operations in the transfer function,
leading to attempts to use a transfer map which had not been initialised.

The reason this is a problem is because setcolorspace stores the newly
created space in the graphics state *before* it builds the lookup table
for the space. The code then puses a 'continuation procedure' to deal
with sampling the transfer function. That procedure pushes the
PostScript procedure and associated data onto the operand stack and then
returns to the interpreter to execute the PostScript. On returning
it stores the returned data and sets up for the next run.

The grestore, however, restores away the colour space which is in the
graphics state, which means that when we come back to the continuation
procedure, it has gone, and attempting to treat the current colour
space as an Indexed space results in accessing invalid memory.

The solution is simply to defer setting the colour space until *after*
we have sampled the function. We do that by storing the colour space on
the execution stack, just as we do with the other parameters, instead
of in the graphics state. We then update the stored colour space, and
only when we have finished sampling the data do we set the colour
space in the graphics state to be the Indexed space.

This means that no matter what the lookup function does, it will be
operating on the state 'before' the space was set, so if it grestores
too many times, it won't matter, the space is safely on the exec stack.
Similarly (shortly) if a function executes a marking operator while
building a transfer function, then the transfer function will be the
one 'prior' to setting the new transfer function, which will get away
from any possibility of trying to use an uninitialised map.

Additional fixes:

PostScript interpreter - defer blackgeneration/undercolorremoval

Defers setting the black generation and undercolorremoval
in the graphics state until *after* we have evaluated the mapping
function.

Settransfer

Added similar code for settransfer, which appears to work.
Checked Separation and DeviceN colour spaces, they do not seem to
suffer the same problem as they convert the PostScript procedure into
a PostScript function before setting the colour space.

setcolortransfer

This seems to work for setcolortransfer. Any function which does weird
stuff (initgraphics, grestore etc) may cause only a partial setup of
the new functions. But then you shouldn't be doing this, the spec
doesn't make it clear what should happen.
Also fixed some compiler warnings.

sethalftone already defers setting the halftone and transfer functions
(if present) until the entire halftone is ready and so does not seem
to suffer this problem.