Bug 706868: Fix stack management when cshow errors
In normal operation, the main interpreter loop maintains a variable (icount)
which is the number of entries still to be consumed in the executable array
currently being executed. At strategic times, when the array is the top object
on the execution stack it uses icount to increment the start of the array to
point to the next object to be executad.

Operators that require a return to the interpreter (such as cshow) use the
multiple entries on the exec stack to store their "private" information as
well as the "continuation function" which the interpreter calls to complete
the operator's function.

The problem here is that cshow hits an error when those exec stack entries are
partially in place. The error is returned to the interpreter loop, and that
triggers that update of the current executable array, which is no longer the
top of the exec stack. That corrupts the top entry on the exec stack, causing
a later crash.

So ensure that when such an error happens, we return the exec stack to its
original state before returning to the interpreter loop.

(this is the same fix used for other members of the "show" family in commit
6c4ef58b66dadf546cbc7370ebbabe5505acc7ed)