Fix compiler warning
I'd included the wrong .h file in zcolor.c. I added gsht.h when I meant
to include gzht.c. This led to a compiler warning.

Also, I forgot to update the int.mak makefile to reflect the fact that
there are new includes in zcolor.c and zcolor1.c so do that now too.