Bug 706892: Handle infinite recursion in error handler
In the even that an error handler procedure causes an error, we can end up in
an infinite recursion, which can end up corrupting either the exec stack or the
operand stack.

Before attempting to execute the error procedure for an execstackoverflow error,
check if we've just tried to execute that procedure. If so, return the error to
caller, rather than try to have Postscript handle it.

Thanks to Alan Goodman for the fuzz testing which uncovered this bug.