Bug 706872: pop_estack() needs to cope with stack blocks
The Ghostscript stacks are extended by adding "blocks", and popping objects
off the stacks can cause a now empty stack block to be freed.

pop_estack() took a pointer reference to a entry on the stack, then popped
entries off down to a special mark objects (including the object it just
referenced). That reference was, effectively, the cleanup function for the
"stack frame" involved.

Problem is, if the popping causes a stack block to be freed, then the ref
containing the cleanup function is also freed.

So save the function pointer, before popping, and use that pointer instead,

Thanks to Alan Goodman for the fuzz testing which uncovered this bug.