Graphics library - fix buffer size for Decode with 16-bit images and ICC
Bug #706866 "heap buffer overflow WRITE decode_row16 / image_render_icc16"

This error appears to date back to the initial commit of this code. The
'decode' buffer was incorrectly sized for the output colour space, but
the Decode array is applied in the source colour space, leading to us
overrunning the buffer.