Bug 706877: Guard against pattern PaintProc doing more saves than restores
Unlike gsave/grestore, we can't reliably recover from this situation, because
of the way that save objects are Postscript objects - unlike gstates which
form an internal stack. So, we can never be sure where to find a given save
state to restore to.

So, record the save level before we run the PaintProc, and check it's the same
on return. If not, throw an error, which will cleanup the state.

Do the same for local/global VM, as an extra security measure.