Bug 706886: Validate array of strings data source for ReusableStreamDecode
Since the array of strings, and the strings themselves, are not (and cannot be
made) read-only for the life of the ReusableStreamDecode, it's possible to
change the contents of the array while the filter is still in use.

On each read, check that we are still dealing with a string of the appropriate
length.