Graphics library - copy source data string for image operator
Bug #706867 "heap use after free READ sample_unpack_8 / gx_image1_plane_data"

The problem is that the specimen file has multiple procedural data
sources; the first executes a save before returning a string, the third
executes a restore back to the save object.

The function gs_image_next_planes() takes a reference to the strings
which are passed in if they are longer than a single image row (shorter
strings are locally copied). This is exposed to the garbage collector
through the image enumerator so ordinarily the strings will persist
until the enumerator has finished with them.

But this does not confer immunity to the restore operator. When the
third procedure executes restore it restores away the two (large)
strings returned by the first two procedures. We then try to use the
retained pointers and fall in a heap.

gs_image_next_planes() gets called with the string data no matter what
the data source is (string, file, procedure) so we must take care not to
break the other cases when fixing this.

The only real solution seems to be to make a local copy of the string
data in 'stable' memory (which is immune to save/restore) and manage
that. Indeed this is what we already do in the case of strings shorter
than one scan line (see the 'row' member of penum->planes).

This means breaking 'const' for the strings but the code already
fiddled with the data and size members, so we weren't honouring the
const anyway (note gs_const_string, not a 'C' const). But it has meant
changing more places in the code to avoid compiler warnings.

I notice that we insist that all the image data source be the same type
which I can't find in the PLRM, but testing Distiller indicates Adobe
make the same restriction.

We do not (contrary to the PLRM) insist that all procedural sources
return the same size string. In fact we can't easily do that because
for type 3 images (explicit mask) the mask and image data sources may
well return different sized strings, the PLRM doesn't say if this is
still a requirement here.