Bug 706876: Fix SEGV with clip_copy_mono and nested imagemasks.
One of the clip_copy_mono operations is dealing with transposed
rectangle data. The optimisation in the clip stuff that holds
the rectangle list data transposed for such cases means that
coords are transposed at various stages, and then get untransposed
later.

The bug here is to do with width and height needing to be passed
untransposed in one case, where they are actually passed transposed.
This looks odd because we appear to be passing x, y transposed, when
actually we are untransposing them!