Bug 706881: Fix SEGV in image_render_mono_ht
The test file sets up an image that is rotated through 270 degrees
before being displayed. This means that we fall into the
right-to-left landscape code.

This means we accumulate data into a landscape buffer before
thresholding it. We start at position 63, and decrement position as
we go. The actual thresholding is not triggered, as you might expect,
by position reaching 0, but by some other mechanism.

Due to the penum->{x,y,w,h}ci values being zero, this mechanism does
not trigger, and so we keep filling in the buffer as position becomes
negative, causing a bufer underrun.

So why are penum->{x,y,w,h}ci not being set? Well, there is some
logic that refuses to set them if interpolate is true. This logic
dates back (with a few changes) to the original 1998 import date.

The landscape thresholding code certainly does not date back that
far! So maybe the assumptions under which those values did not need
to be calculated are now invalid.

It seems reasonable to calculate them even if they are not needed.
They shouldn't do any harm and they are not especially onerous. We
just have to avoid taking the 'mt' exits in the interpolated case.