Graphics library - avoid divide-by-zero during CIE colour space setup
Bug #706874 "SEGV READ cie_lookup_mult3 / gx_cie_xyz_remap_finish"

The problem is actually caused during initialisation of the cache for
the CIEABased colour space. The LMN components cause a divide by zero
(in this case 0/0 so no FPE) which leaves us with an undefined value
in the 'params' member of the cache.

When we try to lookup the cache using an undefined value we end up
trying to read from an invalid address, causing the seg fault.

The CIEBasedA space is clearly nonsense, but not obviously illegal in
any way. So I've chosen to set the (otherwise undefined) values to 0. I
doubt the result will be 'wrong' and it avoids the error.