Bug 706875: Avoid SEGV caused by overflow.
The clip enumerator code manages to avoid a test against NULL
(not that expensive, but it's done a lot) by having a 'stopper'
entry at the end of the list.

Unfortunately, the stopper entry doesn't actually work if
y == max_int on entry. Fix this by just ignoring any calls
with y == ymax. We can worry about this when someone comes
up with a device that large!

Also patch tile_by_steps to avoid us hitting this case with
this test file.