Graphics library - For /Separation > 8-bit images, don't allow fast route
Bug #706880 "SEGV READ gs_image_class_2_fracs / gx_image_enum_begin"

This is obviously an oversight from 16-bit colour processing. Separation
and DeviceN colour spaces with no ICC profile need to run the tint
transform to convert the sample to the alternate space. We were checking
for DeviceN, but not for Separation.