PostScript interpreter - alter passing of functions from Indexed space
Bug #706888 "Heap buffer overflow write of size 2 in ciebasecolor"

The problem is, of course, considerably different from where it
exhibits.

The sample file has an Indexed colour space with a DeviceGray base space
and a tint transform which incorrectly pushes 2 items onto the operand
stack. One is the gray value, which is consumed, the other is the
name '/CIEBasedABC'.

The program then executes a 'settransfer' where the transfer function
executes a currentcmykcolor.

The problem occurs because after completing the Indexed space evaluation
we store the colour space in the interpreter graphics state. We were
assuming that the Indexed lookup function would not disturb the stack
and were simply setting the interpreter state to be whatever was topmost
on the operand stack.

In this case that would be the 'CIEBasedABC' name, rather than the
actual lookup procedure.

We then execute currentcmykcolor, which uses the interpreter state
colour space to decide how to handle the colour returned by
currentcolor, to transform it into the requested colour.

Because the colour space was the name /CIEBasedABC, instead of the
original /Indexed space we would then attempt to remove three colours
from the operand stack, even though only one was pushed. This 'works'
(or at least doesn't crash) by chance, but the second time we do that we
end up running off the bottom of the operand stack and corrupting
memory.

We already store the colour space on the execution stack, so in
setcolorspace_cont() we only need to use the correct element of the
execution stack instead of the operand stack.

However....

For /Separation and DeviceN spaces, if we are able to convert the
tint transform into a function then we replace the PostScript tint
transform with that function (and we store that in the interpreter
state, but it gets overwritten by the original space).

Because we are no longer using the operand stack, we must store the
function on the execution stack (ie replace the ref value on the
execution stack the same way we do on the operand stack).

Finally.... While debugging this I happened to hit a case where the
tint transform conversion to function triggered an exec stack overflow
(Quality Logic file 09-10.ps which leaves > 2000 objects on the operand
stack). This should return to the interpreter loop, increase the stack
if possible and return (or exit with an error). And it does, but because
we had omitted to remove the extra 'setcolorspace_cont' from the
execution stack, all of the stored data was off by one, leading to a
typecheck error. Fixed by checking specifically for a stack overflow in
setcolorspace_cont() and removing the extra continuation before
returning to the interpreter loop.

This (to my complete astonishment) exhibits a progression with the
Quality Logic file 09-10.ps.