PS interpreter - check size of DeviceN ink name before comparing to /All
Bug 706919 "Segfault READ in setdevicenspace (zcolor.c:4319)"

The specimen file has (sort of) a DeviceN colour space with a single
ink name with 0 length. We were checking the ink name bytes before
checking the length which is silly anyway.

Checking the length to see that it is exactly 3 before checking to see
if the name is /All makes more sense and avoids the error with a zero
length string.