PS interpreter - NULL /Indexed map before initialising
Bug #706911 "Segfault READ in gx_final_Indexed (gscolor2.c:300)"

If we get an error when initialising the /Indexed colour lookup we
count down the space we constructed to hold the /Indexed space, and
exit.

But counting down the colour space frees the associated map from the
'params' member. If we hadn't managed to initialise it then we'd try to
free an uninitialised pointer. Just set it to NULL before starting.