PS interpreter & graphics library - check function sanity.
Bug #706910 "Segfault READ in is_tensor_done (gsfunc0.c:633)"

As usual, the actual fault is far from the point where a problem
exhibits.

In this case, due to the excessive number of apparent DeviceN inks (16),
when we create a type 0 (sampled data) function the sampling hypercube
becomes very large, even holding the length of each axis of the
hypercube to 2 (sampling only at 0 and 1). This is, partly, because we
are doing cubic interpolation which causes the required size to balloon
much faster.

First part of the fix is to check that we don't get an overflow when
calculating the required size of the hypercube in gs_function_Sd_init()
if we do then we return a VMerror (makes as much sense as anything else).

This solves the actual reported problem.

The second part of the fix is to try and anticipate and ameliorate the
problem; there is a check in determine_sampled_data_size() which limits
the number of sampled data points to 2 when the number of inks exceeds
8. We now also drop from cubic interpolation to linear interpolation
when that occurs, which reduces the memory overhead making it less
likely we will run into a problem.