PS interpreter - check there is an operand for setcolorspace
Bug #706917 "Segfault READ in ref_stack_counttomark (istack.c:253)"

The specimen file has a /Indexed colour space where the lookup function
also calls setcolorspace with a /Indexed space, and the lookup function
for that space executes a 'clear' which removes the initial lookup
procedure's operand from the stack.

Fix it by checking that we still have at least one operand on the
operand stack before we pop it at the end of setting the space (in this
case the outer colour space).