PS interpreter - avoid infinite recursion
Bug 706956 "Stack overflow in array_get (iutil.c:670)"

The problem is that we have (I think) a 'standard' identity array
'{}' and when we try to recurse into that it gives back the same array.
So we recurse forever.

I think we can assume that if two refs, both of which are arrays, have
the same 'refs' value then they are in fact referencing the same object
and we can short-circuit the test for equality, and avoid the infinite
recursion.