PS interpreter - synchronise effective transfers and real transfers
Bug #706913 "Segfault READ in reloc_struct_ptr (igc.c:1310)"

This is caused by the fact that the 'effective_transfer' pointers in the
graphics state are duplicates of the 'set_stransfer' pointers and as
such are exposed for relocation. But NOT for enumeration.....

The twiddling back and forth of the maps in the graphics state, so that
we can continue to use gs_settransfer_remap() without modification
didn't resynchronise the effecetive_transfer with the actual maps in
set_transfer. This could lead to us trying to relocate a pointer to
memory which had been freed.

Here we just call gx_set_effective_transfer() whenever we change one of
the transfer map pointers. This resolves the problem as the two sets
of pointers are always synchronised.