Bug 706922: Fix filenameforall completion cleanup
Previously, zfilenameforall() relied upon the specifics of normal interpreter
loop operation to correctly cleanup on completion, which works unless the
interpreter loop is interrupted by a garbage collection which fails (VMerror).
In which case, the top objects on the exec stack are not correctly shuffled
around, and the file_cleanup() function ends up being called erroneously with
a corrupted file enumerator object which has already been freed anyway.

So simply don't use the "trick" and cleanup the stack explicitly before we
return to the interpreter loop.