PS interpreter - fix problem with exec stack pointer and forall
Bug #706938 "Segfault WRITE in zcolor_remap_one_store (zcolor.c:805)"

In fact the problem is not really to do with the mapping. The problem is
that the under color removal function calls currentsystemparams, which
is defined in PostScript and uses forall to enumerate a dictionary.

The actual problem occurs during forall; We make a pointer to the exec
stack pointer, and define an object based on that, then we call
check_estack(). That discovers there is not enough space on the exec
stack so it copies it to a new larger block. This, of course,
invalidates the existing pointers. We then use those pointers and end
up corrupting the exec stack.

Simply move the pointer assignment after the check of the stack.