graphics library - prevent negative memcpy count
bug #706957 "Negative-size-param in fill_threshold_buffer (gxht_thresh.c:788)"

I can't quite see how this is possible, or what it means in reality.
However given that we don't have anyone with a better clue than me, the
simplest solution seems to be just to check and prevent a copy of a
negative number of bytes. Copying 0 bytes has no effect so can#'t be
a problem.