PS interpreter - range check arguments to trigonometric extensions
Bug #706974 "Segfault READ in is_tensor_done (gsfunc0.c:633)"

As usual the actual fault lies far from the seg fault. The actual
problem is that the extension PostScript operators for some
trigonometric maths functions were not checking their argument. This led
to an 'undefined' number on the operand stack and the consequent
attempt to use that as a parameter to setcolor causes the crash.

This commit range checks the arccos and arcsin operators, as well as
the 'double precision' variants.